move k8s clientset into helper method, use client go to retrieve k8s version
